projects: also move ProjectBanner to Views/projects/banner
Folded into this PR per follow-up review: `ProjectBanner` belongs
at `Views::Controllers::Projects::Banner` for the same reason as
the other Project-focused views moved here. The cross-cluster
caller (`app/views/controllers/checklists/show.rb`) references it
via the full namespace.

  Components::ProjectBanner → Views::Controllers::Projects::Banner

Updated the `add_project_banner` helper in
`app/helpers/tabs/projects_helper.rb` to reference the new
namespace. Inside the new file, the sibling `Tabs` is referenced
by lexical-scope lookup (`render(Tabs.new(...))`) — same shape as
the other moved projects classes.

19 banner tests + 290 projects-cluster controller tests pass.
Lint clean.
